in preparation for the new asset manager page, exported the whole explore page into its own component, AssetExplorer, with localstorage capabilities and refactored the Explore page to use it
